아이템 소지수/중첩, 인벤토리 창 구현, 변수 타입 정리
인벤토리 씬으로 인벤토리 구현
 - 입력제어에 용이하도록 씬으로 구현

아이템 소지 개수, 중첩 처리

배열 index, .size() 에 사용되는 것들은 size_t 로 변경 (역순회 시엔 int 사용)
값이 항상 양수인것은 unsigned int로 변경
계산 시 음수가 될 수 있는것들은 int 유지

+++ BaseGame/BattleManager.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "common.h"
+#include <vector>
+#include <memory>
 
 class Player;
 class Monster;
@@ -12,20 +13,20 @@ public:
 	~BattleManager() = default;
 
 	void StartBattle(const std::vector<Monster*>& monsters);
-	void PlayerAttack(int target_idx);
+	void PlayerAttack(size_t target_idx);
 	void MonsterAttack();
 
 	bool IsBattleOver() const;
 	bool IsPlayerVictory() const;
 	void DistributeReward();	// 전투 종료시 보상 지급
 
-	int GetTotalExp() const;
+	unsigned int GetTotalExp() const;
 
 private:
 	Player* player = nullptr;
 	std::vector<Monster*> monster_list;
-	int total_exp = 0;
-	int total_gold = 0;
+	unsigned int total_exp = 0;
+	unsigned int total_gold = 0;
 	std::vector<std::unique_ptr<Item>> items;
 };
